refactored navigation stuff to separate folder, excluded logo and title as separate component
