+++ core/src/navigation/LeftNav.html
@@ -53,10 +53,10 @@
 </div>
 
 <script type="text/javascript">
-    import Backdrop from './Backdrop.html';
+    import Backdrop from '../Backdrop.html';
     import { getChildren, getLeftNavData } from './services/navigation.js';
-    import { handleRouteClick } from './services/routing.js';
-    import { getConfigValue, getConfigBooleanValue } from './services/config.js';
+    import { handleRouteClick } from '../services/routing.js';
+    import { getConfigValue, getConfigBooleanValue } from '../services/config.js';
 
     const setLeftNavData = async (current, component) => {
       const componentData = component.get();
