+++ core/src/navigation/LogoTitle.html
@@ -1,0 +1,63 @@
+
+<div on:click="goTo('/')" class="fd-global-nav__logo fd-has-margin-left-none" ref:logo style="background-image: 'none'"></div>
+<div class="fd-global-nav__product-name">
+  {#if title}
+  <a on:click="goTo('/')">{title}</a>
+  {/if}
+</div>
+
+<script type="text/javascript">
+  import { getConfigValueAsync, getConfigBooleanValue } from '../services/config.js';
+
+  export default {
+    oncreate() {
+      getConfigValueAsync('settings.header').then((header) => {
+        if (!header) { return; }
+
+        // Set Title and Logo
+        if (header.title) {
+          this.set({ title: header.title });
+          document.title = header.title;
+        }
+        if (header.logo) {
+          this.refs.logo.style.backgroundImage = 'url(' + header.logo + ')';
+        }
+
+        // Set Favicon
+        if (header.favicon) {
+          if (!header.favicon.split('?')[0].endsWith('.ico')) {
+            console.warn('Favicon is not an .ico filetype and might get displayed wrong.');
+          }
+          const link = document.createElement('link');
+          link.type = 'image/x-icon';
+          link.rel = 'shortcut icon';
+          link.href = header.favicon;
+          document.getElementsByTagName('head')[0].appendChild(link);
+        }
+      });
+    },
+    methods: {
+      goTo(path) {
+        if (getConfigBooleanValue('routing.useHashRouting')) {
+          window.location.hash = path;
+        } else {
+          window.location.pathname = path;
+        }
+      }
+    }
+  };
+</script>
+
+<style type="text/scss">
+  @import 'node_modules/fundamental-ui/scss/components/global-nav';
+  .fd-global-nav__logo {
+    cursor: pointer;
+  }
+  .fd-global-nav__product-name {
+    color: #0a6ed1;
+    height: 50px;
+    font-size: 24px;
+    line-height: 1em;
+    padding: 13px 0;
+  }
+</style>
